I have implemented forward-search for the PdfSync module. SumatraPDF now accepts DDE queries of the form:
[ForwardSearch("c:\file.pdf","c:\folder\source.tex",298)]

This tells Sumatra to go to the location in 'c:\file.pdf' that corresponds to the line 298 of the source file 'c:\folder\source.tex'.

Somehow it only works in continuous mode because the function win->dm->goToPage that I use to move inside the PDF document does not work properly in single page mode.

git-svn-id: http://sumatrapdf.googlecode.com/svn/trunk@685 841c7e00-0c1d-0410-a694-8ba26473ccb1